requirements.txt
python-dotenv
pymongo[srv]
pandas
requests
selenium
webdriver-manager
fastapi
uvicorn
openai